Moving Console\Error to Console\Exception.
Also fixed risky test
