+++ src/Console/Exception/ConsoleException.php
@@ -11,16 +11,14 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Console\Error;
+namespace Cake\Console\Exception;
 
 use Cake\Core\Exception\Exception;
 
 /**
- * Used when a shell method cannot be found.
+ * Exception class for Console libraries.  This exception will be thrown from Console library
+ * classes when they encounter an error.
  *
  */
-class MissingShellMethodException extends Exception {
-
-	protected $_messageTemplate = "Unknown command %1\$s %2\$s.\nFor usage try `cake %1\$s --help`";
-
+class ConsoleException extends Exception {
 }
